Split count_distinct.rs into separate modules  (#9087)
* Split count_distinct.rs into separate modules

* Remove unecessary typedef

* Rename

* improve module comments
